@@ -3,30 +3,30 @@
     "version": "0.2.4",
     "private": true,
     "dependencies": {
-        "@emotion/react": "^11.4.0",
-        "@emotion/styled": "^11.3.0",
+        "@emotion/react": "^11.8.2",
+        "@emotion/styled": "^11.8.1",
         "@material-ui/core": "^5.0.0-beta.5",
         "@material-ui/data-grid": "^4.0.0-alpha.37",
         "@material-ui/icons": "^5.0.0-beta.5",
         "@material-ui/styles": "^5.0.0-beta.5",
-        "@testing-library/jest-dom": "^5.14.1",
-        "@testing-library/react": "^11.2.7",
-        "@testing-library/user-event": "^12.8.3",
-        "@types/jest": "^26.0.24",
-        "@types/node": "^12.20.17",
-        "@types/react": "^17.0.15",
-        "@types/react-dom": "^17.0.9",
-        "@types/react-router-dom": "^5.1.8",
+        "@testing-library/jest-dom": "^5.16.2",
+        "@testing-library/react": "^12.1.4",
+        "@testing-library/user-event": "^13.5.0",
+        "@types/jest": "^27.4.1",
+        "@types/node": "^17.0.21",
+        "@types/react": "^17.0.40",
+        "@types/react-dom": "^17.0.13",
+        "@types/react-router-dom": "^5.3.3",
         "chart.js": "^3.5.1",
-        "lorem-ipsum": "^2.0.3",
+        "lorem-ipsum": "^2.0.4",
         "react": "^17.0.2",
         "react-chartjs-2": "^3.0.4",
         "react-dom": "^17.0.2",
         "react-helmet": "^6.1.0",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "4.0.3",
-        "typescript": "^4.4.2",
-        "web-vitals": "^1.1.2"
+        "react-router-dom": "^6.2.2",
+        "react-scripts": "^5.0.0",
+        "typescript": "^4.6.2",
+        "web-vitals": "^2.1.4"
     },
     "scripts": {
         "start": "react-scripts start",
@@ -59,10 +59,10 @@
         ]
     },
     "devDependencies": {
-        "@types/react-helmet": "^6.1.2",
-        "@typescript-eslint/eslint-plugin": "^5.12.0",
-        "eslint": "^8.9.0",
-        "eslint-config-prettier": "^8.3.0",
+        "@types/react-helmet": "^6.1.5",
+        "@typescript-eslint/eslint-plugin": "^5.14.0",
+        "eslint": "^8.10.0",
+        "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-prettier": "^4.0.0",
         "prettier": "^2.5.1"
     }
