@@ -10,20 +10,14 @@ import {
     Typography,
 } from "@material-ui/core";
 import { styled } from "@material-ui/core/styles";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import { NotFoundPage } from "./pages/NotFound";
+import { BrowserRouter } from "react-router-dom";
 import { Link } from "./components";
 import InvertColorsIcon from "@material-ui/icons/InvertColors";
 import { useThemeToggle } from "./Theming";
-import { LogInPage } from "./pages/LogIn";
 import { UserInfo } from "./components/UserInfo";
 import { Footer, FOOTER_HEIGHT } from "./components/Footer";
-import { PrivacyPage } from "./pages/Privacy";
-
-const HomePage = React.lazy(() => import("./pages/Home"));
-const RaiderPage = React.lazy(() => import("./pages/Raider"));
-const RaidTeamsPage = React.lazy(() => import("./pages/RaidTeams"));
-const RaidTeamPage = React.lazy(() => import("./pages/RaidTeam"));
+import { ErrorBoundary } from "./components/ErrorBoundary";
+import { MainRouting } from "./Routing";
 
 const Main = styled("main")(({ theme }) => ({
     flexGrow: 1,
@@ -47,27 +41,29 @@ const ToolbarRootLink = styled(Link)(() => ({
     gap: 20,
 }));
 
-const ToolbarTitle = styled(Typography)(() => ({
+const ToolbarTitle = styled("div")(() => ({
     padding: "1.25rem 0",
+    display: "flex",
+    flexDirection: "row",
 }));
 
 export function App() {
     return (
-        <Router>
+        <BrowserRouter>
             <CssBaseline />
             <Box sx={{ display: "flex", flexDirection: "column" }}>
                 <Header />
                 <Main>
                     <Toolbar />
                     <React.Suspense fallback={<PageLoading />}>
-                        <SwitchAndRoutes />
+                        <MainRouting />
                     </React.Suspense>
                 </Main>
                 <footer>
                     <Footer />
                 </footer>
             </Box>
-        </Router>
+        </BrowserRouter>
     );
 }
 
@@ -78,64 +74,39 @@ function Header() {
     return (
         <AppBar position="fixed">
             <Toolbar>
-                <ToolbarRootLink to="/">
-                    <img
-                        alt="RaidSIMP Icon"
-                        src="/eye_logo.png"
-                        style={{
-                            objectFit: "contain",
-                            maxHeight: 64,
-                        }}
-                    />
-                    <ToolbarTitle variant="h5" color="textPrimary" gutterBottom={false}>
-                        RaidSIMP
-                    </ToolbarTitle>
-                </ToolbarRootLink>
+                <ErrorBoundary>
+                    <ToolbarRootLink to="/">
+                        <img
+                            alt="RaidSIMP Icon"
+                            src="/eye_logo.png"
+                            style={{
+                                objectFit: "contain",
+                                maxHeight: 64,
+                            }}
+                        />
+                        <ToolbarTitle>
+                            <Typography variant="h5" color="textPrimary">
+                                Raid
+                            </Typography>
+                            <Typography variant="h5" color="secondary">
+                                SIMP
+                            </Typography>
+                        </ToolbarTitle>
+                    </ToolbarRootLink>
 
-                <Stack direction={"row"} alignItems={"center"} spacing={3}>
-                    {showThemeToggle && (
-                        <IconButton onClick={toggleTheme} color="inherit" title="Toggle theme">
-                            <InvertColorsIcon />
-                        </IconButton>
-                    )}
-                    <React.Suspense fallback={<></>}>
-                        <UserInfo />
-                    </React.Suspense>
-                </Stack>
+                    <Stack direction={"row"} alignItems={"center"} spacing={3}>
+                        {showThemeToggle && (
+                            <IconButton onClick={toggleTheme} color="inherit" title="Toggle theme">
+                                <InvertColorsIcon />
+                            </IconButton>
+                        )}
+                        <React.Suspense fallback={<></>}>
+                            <UserInfo />
+                        </React.Suspense>
+                    </Stack>
+                </ErrorBoundary>
             </Toolbar>
         </AppBar>
-    );
-}
-
-function SwitchAndRoutes() {
-    return (
-        <Switch>
-            <Route exact path="/">
-                <HomePage />
-            </Route>
-            <Route
-                path="/raid-teams/:teamId/raiders/:raiderId"
-                render={({ match }) => (
-                    <RaiderPage teamId={match.params.teamId} raiderId={match.params.raiderId} />
-                )}
-            />
-            <Route
-                path="/raid-teams/:teamId"
-                render={({ match }) => <RaidTeamPage teamId={match.params.teamId} />}
-            />
-            <Route path="/raid-teams">
-                <RaidTeamsPage />
-            </Route>
-            <Route path="/login">
-                <LogInPage />
-            </Route>
-            <Route path="/privacy">
-                <PrivacyPage />
-            </Route>
-            <Route path="*">
-                <NotFoundPage />
-            </Route>
-        </Switch>
     );
 }
 
