@@ -1,4 +1,5 @@
 import { Container, Link, Paper, styled, Tooltip, Typography } from "@material-ui/core";
+import { ErrorBoundary } from "./ErrorBoundary";
 
 export const FOOTER_HEIGHT = "72px";
 
@@ -18,30 +19,36 @@ const FootieLinkContainer = styled(Container)(() => ({
 export function Footer() {
     return (
         <FootiePaper elevation={12} square>
-            <FootieLinkContainer>
-                <Link href="mailto:raidsimpapp@gmail.com" target="_blank" rel="noopener noreferrer">
-                    <Typography>Contact us</Typography>
-                </Link>
-                <Link href="https://twitter.com/LinusPhoenix">
-                    <Typography>Tweet at us</Typography>
-                </Link>
-                <Tooltip disableInteractive placement="top" title="Coming Soon!">
-                    <Link href="#">
-                        <Typography>Join us on Discord</Typography>
+            <ErrorBoundary>
+                <FootieLinkContainer>
+                    <Link
+                        href="mailto:raidsimpapp@gmail.com"
+                        target="_blank"
+                        rel="noopener noreferrer"
+                    >
+                        <Typography>Contact us</Typography>
                     </Link>
-                </Tooltip>
-                <Link href="https://github.com/LinusPhoenix/wow-raid-manager">
-                    <Typography>View our code on GitHub</Typography>
-                </Link>
-                <Tooltip disableInteractive placement="top" title="Coming Soon!">
-                    <Link href="#">
-                        <Typography>Support us on Ko-fi</Typography>
+                    <Link href="https://twitter.com/LinusPhoenix">
+                        <Typography>Tweet at us</Typography>
                     </Link>
-                </Tooltip>
-                <Link href="/privacy">
-                    <Typography>Privacy Policy</Typography>
-                </Link>
-            </FootieLinkContainer>
+                    <Tooltip disableInteractive placement="top" title="Coming Soon!">
+                        <Link href="#">
+                            <Typography>Join us on Discord</Typography>
+                        </Link>
+                    </Tooltip>
+                    <Link href="https://github.com/LinusPhoenix/wow-raid-manager">
+                        <Typography>View our code on GitHub</Typography>
+                    </Link>
+                    <Tooltip disableInteractive placement="top" title="Coming Soon!">
+                        <Link href="#">
+                            <Typography>Support us on Ko-fi</Typography>
+                        </Link>
+                    </Tooltip>
+                    <Link href="/privacy">
+                        <Typography>Privacy Policy</Typography>
+                    </Link>
+                </FootieLinkContainer>
+            </ErrorBoundary>
         </FootiePaper>
     );
 }
