@@ -3,13 +3,19 @@ import ReactDOM from "react-dom";
 import { App } from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { ThemeProvider } from "./Theming";
+import { UserInfoProvider } from "./components/UserInfoContext";
 import { PromiseProvider } from "./utility";
+import { ErrorBoundary } from "./components/ErrorBoundary";
 
 ReactDOM.render(
     <React.StrictMode>
         <ThemeProvider>
             <PromiseProvider>
-                <App />
+                <UserInfoProvider>
+                    <ErrorBoundary>
+                        <App />
+                    </ErrorBoundary>
+                </UserInfoProvider>
             </PromiseProvider>
         </ThemeProvider>
     </React.StrictMode>,
